Created Joke Name Generator service
Service takes GET requests and responds with a joke string.
The joke contains a name (sourced from uinames.com) inserted into
a joke from the Chuck Norris database (icndb.com).
The project uses gorilla/mux to handle and route requests.
Basic concurrency is employed to make input request to the two
dependent services.
Service-level logging is provided to stdout.
The project uses Go Modules for dependency management.
A description of the work completed, to be completed, and
instructions on how to use the service are in the README.